% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{aFiles}
\alias{aFiles}
\alias{aFiles<-}
\alias{aFiles,ISAjson-method}
\title{Get and set aFiles.}
\usage{
aFiles(x)

aFiles(x) <- value

\S4method{aFiles}{ISAjson}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the assay files.
In each data.frame at least the following columns are present:
`r paste0("'", aFilesCols, "'", collapse = ", ")`.}
}
\value{
A list of data.frames containing the assay files.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of aFiles data.frames in an object of
\linkS4class{ISAjson}.
}
