% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{aMaterials}
\alias{aMaterials}
\alias{aMaterials<-}
\alias{aMaterials,ISAjson-method}
\title{Get and set aMaterials}
\usage{
aMaterials(x)

aMaterials(x) <- value

\S4method{aMaterials}{ISAjson}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the assay materials.}
}
\value{
A list of data.frames containing the assay materials.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of aMaterials data.frames in an object of
\linkS4class{ISAjson}.
}
