% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{aUnitCats}
\alias{aUnitCats}
\alias{aUnitCats<-}
\alias{aUnitCats,ISAjson-method}
\title{Get and set aUnitCats.}
\usage{
aUnitCats(x)

aUnitCats(x) <- value

\S4method{aUnitCats}{ISAjson}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the assay unit categories.}
}
\value{
A list of data.frames containing the assay unit categories.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of aUnitCats data.frames in an object of
\linkS4class{ISAjson}.
}
