% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setInfo.R
\name{getStudyFileNames}
\alias{getStudyFileNames}
\title{Retrieve the Study File Name(s) and Study File Identifier(s)from an ISAjson
object.}
\usage{
getStudyFileNames(isaObject)
}
\arguments{
\item{isaObject}{An object of the \linkS4class{ISAjson}.}
}
\value{
A named character vector containing the Study File Name(s) and the
name(s) representing the Study Identifier(s).
}
\description{
Retrieve from an object of the \linkS4class{ISAjson} the Study
Identifier(s) and Study File Name(s) as contained in the Investigation.
To directly access the Study Identifier(s) use the names() function, e.g.
\code{names(getStudyFileNames(isaObject))}.
}
