% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{iContacts}
\alias{iContacts}
\alias{iContacts<-}
\alias{iContacts,ISAjson-method}
\alias{iContacts<-,ISAjson-method}
\title{Get and set iContacts.}
\usage{
iContacts(x)

iContacts(x) <- value

\S4method{iContacts}{ISAjson}(x)

\S4method{iContacts}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A data.frame containing the investigation contacts
information. A data.frame in which at least the following columns are
present:
`r paste0("'", iContactsCols, "'", collapse = ", ")`.}
}
\value{
A data.frame containing the investigation contacts information.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the iContacts data.frame in an object of
\linkS4class{ISAjson}.
}
