% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{invest}
\alias{invest}
\alias{invest<-}
\alias{invest,ISAjson-method}
\alias{invest<-,ISAjson-method}
\title{Get and set invest.}
\usage{
invest(x)

invest(x) <- value

\S4method{invest}{ISAjson}(x)

\S4method{invest}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A data.frame containing the investigation information. A
data.frame in which at least the following columns are present:
`r paste0("'", investCols, "'", collapse = ", ")`.}
}
\value{
A data.frame containing the investigation information.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the investigation data.frame in an object of
\linkS4class{ISAjson}.
}
