% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{iPubs}
\alias{iPubs}
\alias{iPubs<-}
\alias{iPubs,ISAjson-method}
\alias{iPubs<-,ISAjson-method}
\title{Get and set iPubs.}
\usage{
iPubs(x)

iPubs(x) <- value

\S4method{iPubs}{ISAjson}(x)

\S4method{iPubs}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A data.frame containing the investigation publications
information. A data.frame in which at least the following columns are
present:
`r paste0("'", pubsCols, "'", collapse = ", ")`.}
}
\value{
A data.frame containing the investigation publications information.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the iPubs data.frame in an object of \linkS4class{ISAjson}.
}
