% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{ISAjson-class}
\alias{ISAjson-class}
\alias{ISAjson}
\title{S4 Class ISAjson, initialization method}
\description{
An S4 class to store information from an ISA-JSON file.
}
\section{Slots}{

\describe{
\item{\code{path}}{A length-one character vector containing the path to the ISA-JSON
file.}

\item{\code{content}}{The file content.}
}}

\keyword{classes}
