% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{initialize,ISAjson-method}
\alias{initialize,ISAjson-method}
\alias{ISAjson-initialize}
\title{Constructor method of ISAjson-class.}
\usage{
\S4method{initialize}{ISAjson}(.Object, path)
}
\arguments{
\item{.Object}{character, name of the object of class \linkS4class{ISAjson}
to be initialized}

\item{path}{length-one character vector containing the path to the ISA-JSON
file.}
}
\description{
When creating a new object of class \linkS4class{ISAjson} via:
\code{object <- new(Class = "ISAjson", path)}, the function
\code{initialize(.Object, path)} is called to initialize and create the
actual object. The \code{initialize-method} is seldom used as a function
itself.
}
