% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isajsonr.R
\docType{package}
\name{isajsonr-package}
\alias{isajsonr}
\alias{isajsonr-package}
\title{isajsonr package overview.}
\description{
The isajsonr package is an implementation of the ISA-JSON format for R. It
builds an R object from ISA-JSON format files.\cr\cr
ISA is a metadata framework to manage an increasingly diverse set of life
science, environmental and biomedical experiments that employ one or a
combination of technologies. Built around the **Investigation** (the
project context), **Study** (a unit of research) and **Assay**
(analytical measurements) concepts, ISA helps you to provide rich
descriptions of experimental metadata (i.e. sample characteristics,
technology and measurement types, sample-to-data relationships) so that the
resulting data and discoveries are reproducible and reusable.
}
\details{
The ISA Abstract Model has been implemented in two format specifications,
ISA-Tab and ISA-JSON. In this package the latter, ISA-JSON, is being used.
ISA-JSON files are json files stored with a **.json** extension.
}
\references{
[ISA framework](https://isa-tools.org/)
[ISA Model and Serialization Specifications](https://isa-specs.readthedocs.io/en/stable/)
[ISA users community](https://www.isacommons.org)
}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl} (\href{https://orcid.org/0000-0002-8673-2514}{ORCID})

}
\keyword{internal}
