% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{oSR}
\alias{oSR}
\alias{oSR<-}
\alias{oSR,ISAjson-method}
\alias{oSR<-,ISAjson-method}
\title{Get and set oSR}
\usage{
oSR(x)

oSR(x) <- value

\S4method{oSR}{ISAjson}(x)

\S4method{oSR}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A data.frame containing the ontology Source Reference
information. data.frame in which at least the following columns are present:
`r paste0("'", oSRCols, "'", collapse = ", ")`.}
}
\value{
A data.frame containing the ontology Source Reference information.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the ontology Source Reference (oSR) data.frame in an object of
\linkS4class{ISAjson}.
}
