% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readISAJSON}
\alias{readISAJSON}
\title{Read an ISA-JSON file into an R object.}
\usage{
readISAJSON(file, zipfile = NULL, verbose = FALSE)
}
\arguments{
\item{file}{A character string with the name of the ISA-JSON file is located
(if the parameter zipfile is not provided or if it
is equal to \code{NULL}), or the name of the directory where the zip archive
containing the ISA-JSON file is located (if the parameter zipfile is not
\code{NULL}).}

\item{zipfile}{A character vector with the name of the zip archive containing
ISA-JSON file itself (without a directory name in the zip archive). The
default value is \code{NULL} (specifying that the ISA-JSON file has not
been archived in a zipped file).}

\item{verbose}{Should the messages for the different reading steps be shown?}
}
\value{
An object of the \linkS4class{ISAjson} class.
}
\description{
Reads an ISA-JSON file, given as a zip file or as a json file in a
specific folder, and builds an object of the \linkS4class{ISAjson} class.
}
