% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sAssays}
\alias{sAssays}
\alias{sAssays<-}
\alias{sAssays,ISAjson-method}
\alias{sAssays<-,ISAjson-method}
\title{Get and set sAssays.}
\usage{
sAssays(x)

sAssays(x) <- value

\S4method{sAssays}{ISAjson}(x)

\S4method{sAssays}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the study assays.
In each data.frame at least the following columns are present:
`r paste0("'", sAssaysCols, "'", collapse = ", ")`.}
}
\value{
A list of data.frames containing the study assays.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of sAssays data.frames in an object of
\linkS4class{ISAjson}.
}
