% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sCharCats}
\alias{sCharCats}
\alias{sCharCats<-}
\alias{sCharCats,ISAjson-method}
\alias{sCharCats<-,ISAjson-method}
\title{Get and set sCharCats.}
\usage{
sCharCats(x)

sCharCats(x) <- value

\S4method{sCharCats}{ISAjson}(x)

\S4method{sCharCats}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the study characteristic
categories. In each data.frame at least the following columns are present:
`r paste0("'", sCharCatsCols, "'", collapse = ", ")`.}
}
\value{
A list of data.frames containing the study characteristic categories.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of sCharCats data.frames in an object of
\linkS4class{ISAjson}.
}
