% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sDD}
\alias{sDD}
\alias{sDD<-}
\alias{sDD,ISAjson-method}
\alias{sDD<-,ISAjson-method}
\title{Get and set sDD.}
\usage{
sDD(x)

sDD(x) <- value

\S4method{sDD}{ISAjson}(x)

\S4method{sDD}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the study design descriptors.
In each data.frame at least the following columns are present:
`r paste0("'", sDDCols, "'", collapse = ", ")`.}
}
\value{
A list of data.frames containing the study design descriptors.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of sDD data.frames in an object of
\linkS4class{ISAjson}.
}
