% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sFacts}
\alias{sFacts}
\alias{sFacts<-}
\alias{sFacts,ISAjson-method}
\alias{sFacts<-,ISAjson-method}
\title{Get and set sFacts.}
\usage{
sFacts(x)

sFacts(x) <- value

\S4method{sFacts}{ISAjson}(x)

\S4method{sFacts}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISA}.}

\item{value}{A list of data.frames containing the study factors.
In each data.frame at least the following columns are present:
`r paste0("'", sFactsCols, "'", collapse = ", ")`.}
}
\value{
A list of data.frames containing the study factors.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of sFacts data.frames in an object of
\linkS4class{ISAjson}.
}
