% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sMaterials}
\alias{sMaterials}
\alias{sMaterials<-}
\alias{sMaterials,ISAjson-method}
\title{Get and set sMaterials}
\usage{
sMaterials(x)

sMaterials(x) <- value

\S4method{sMaterials}{ISAjson}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the study materials.}
}
\value{
A list of data.frames containing the study protocols.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of sMaterials data.frames in an object of
\linkS4class{ISAjson}.
}
