% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sProcSeq}
\alias{sProcSeq}
\alias{sProcSeq<-}
\alias{sProcSeq,ISAjson-method}
\alias{sProcSeq<-,ISAjson-method}
\title{Get and set sProcSeq.}
\usage{
sProcSeq(x)

sProcSeq(x) <- value

\S4method{sProcSeq}{ISAjson}(x)

\S4method{sProcSeq}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the study process sequences.
In each data.frame at least the following columns are present:
`r paste0("'", sProcSeqCols, "'", collapse = ", ")`.}
}
\value{
A list of data.frames containing the study process sequences

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of sProcSeq data.frames in an object of
\linkS4class{ISAjson}.
}
