% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sProts}
\alias{sProts}
\alias{sProts<-}
\alias{sProts,ISAjson-method}
\alias{sProts<-,ISAjson-method}
\title{Get and set sProts.}
\usage{
sProts(x)

sProts(x) <- value

\S4method{sProts}{ISAjson}(x)

\S4method{sProts}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the study protocols.
In each data.frame at least the following columns are present:
`r paste0("'", sProtsCols, "'", collapse = ", ")`.}
}
\value{
A list of data.frames containing the study protocols.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of sProts data.frames in an object of
\linkS4class{ISAjson}.
}
