% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sPubs}
\alias{sPubs}
\alias{sPubs<-}
\alias{sPubs,ISAjson-method}
\alias{sPubs<-,ISAjson-method}
\title{Get and set sPubs.}
\usage{
sPubs(x)

sPubs(x) <- value

\S4method{sPubs}{ISAjson}(x)

\S4method{sPubs}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the study publications.
In each data.frame at least the following columns are present:
`r paste0("'", pubsCols, "'", collapse = ", ")`.}
}
\value{
A list of data.frames containing the study publications.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of sPubs data.frames in an object of
\linkS4class{ISAjson}.
}
