% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{sUnitCats}
\alias{sUnitCats}
\alias{sUnitCats<-}
\alias{sUnitCats,ISAjson-method}
\alias{sUnitCats<-,ISAjson-method}
\title{Get and set sUnitCats.}
\usage{
sUnitCats(x)

sUnitCats(x) <- value

\S4method{sUnitCats}{ISAjson}(x)

\S4method{sUnitCats}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A list of data.frames containing the study unit categories.}
}
\value{
A list of data.frames containing the study unit categories.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the list of sUnitCats data.frames in an object of
\linkS4class{ISAjson}.
}
