% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validity.R
\name{validISAJSONObject}
\alias{validISAJSONObject}
\title{Check the validity of an object of class ISAjson.}
\usage{
validISAJSONObject(object)
}
\arguments{
\item{object}{An object of class \linkS4class{ISAjson}.}
}
\value{
TRUE or an error message.
}
\description{
The \code{validISAObject} function checks whether an object of class
\linkS4class{ISAjson} is a valid object. An object of the
\linkS4class{ISAjson} is considered valid when:
* The structure of the json file matches the schema for a valid ISA-JSON
file that can be found at
https://github.com/ISA-tools/isa-api/tree/master/isatools/resources/schemas/isa_model_version_1_0_schemas/core
*
}
\seealso{
\linkS4class{ISAjson}
}
