% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeISAjson}
\alias{writeISAjson}
\title{Write ISA-JSON object.}
\usage{
writeISAjson(isaObject, path = getwd())
}
\arguments{
\item{isaObject}{An object of the \linkS4class{ISAjson}.}

\item{path}{A character vector with the name of the directory to which the
file(s) should be written. The default value is the current working
directory.}
}
\value{
No return value, files are written to path.
}
\description{
Write ISA-JSON object to file. The file is written to the folder specified in
\code{path}.
}
