% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/setInfo.R
\name{isaPath}
\alias{isaPath}
\alias{isaPath<-}
\alias{isaPath,ISAjson-method}
\alias{isaPath<-,ISAjson-method}
\title{Get and set path}
\usage{
isaPath(x)

isaPath(x) <- value

\S4method{isaPath}{ISAjson}(x)

\S4method{isaPath}{ISAjson}(x) <- value
}
\arguments{
\item{x}{An object of class \linkS4class{ISAjson}.}

\item{value}{A length-one character vector indicating the isaPath to an
accessible directory on the system.}
}
\value{
The path to the folder for the object of class \linkS4class{ISAjson}.

The updated object of class \linkS4class{ISAjson}.
}
\description{
Get and set the file path for an object of class \linkS4class{ISAjson}.
}
