% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAssay.R
\docType{methods}
\name{processAllAssays}
\alias{processAllAssays}
\title{Process all assay tab data}
\usage{
processAllAssays(isaObject, study, type = c("raw", "derived"))
}
\arguments{
\item{isaObject}{An object of the \linkS4class{ISAjson}.}

\item{type}{A character string indicating which data files should be
processed, either "raw" for raw data files, or "derived" for derived data
files. The file names are taken from the corresponding column in the
\code{aTabObject}s.}
}
\description{
Process data from all assay tab files present in an object of
\linkS4class{ISAjson}.
}
