% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processAssay.R
\docType{methods}
\name{processAssay}
\alias{processAssay}
\alias{processAssay,ISAjson,assayTab,character-method}
\alias{processAssay,ISA,assayTab-method}
\alias{processAssay,ISAjson,msAssayTab,character-method}
\alias{processAssay,ISA,msAssayTab-method}
\alias{processAssay,ISAjson,microarrayAssayTab,character-method}
\title{Process assay tab data}
\usage{
processAssay(isaObject, aTabObject, type = c("raw", "derived"))

\S4method{processAssay}{ISAjson,assayTab,character}(isaObject, aTabObject, type = c("raw", "derived"))

\S4method{processAssay}{ISAjson,msAssayTab,character}(isaObject, aTabObject, type = c("raw", "derived"))

\S4method{processAssay}{ISAjson,microarrayAssayTab,character}(isaObject, aTabObject, type = c("raw", "derived"))
}
\arguments{
\item{isaObject}{An object of the \linkS4class{ISAjson}.}

\item{aTabObject}{An object of the \linkS4class{microarrayAssayTab}.}

\item{type}{A character string indicating which data files should be
processed, either "raw" for raw data files, or "derived" for derived data
files. The file names are taken from the corresponding column in the
\code{aTabObject}.}
}
\description{
Process data from assay tab file

Process data from assay tab files with technology type mass spectrometry
(ms). Processing those files requires the xcms package to be installed.

Process data from assay tab files with technology type DNA microarray
(ms). Processing those files requires the Biobase and affy packages to be
installed.
}
