% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{writeISAjson}
\alias{writeISAjson}
\title{Write ISA-JSON object.}
\usage{
writeISAjson(isaObject, file)
}
\arguments{
\item{isaObject}{An object of the \linkS4class{ISAjson}.}

\item{file}{A character string with the name of the file that should be
written.}
}
\value{
No return value, file is written.
}
\description{
Write ISA-JSON object to file.
}
